+++ media/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Call Stack Ninja</title>
+</head>
+<body>  
+    <h3>Hello From Call Stack Ninja!</h3>
+    <h5>We make call stack easy to visualize</h5>
+    <p>You only need to add breakpoints to see the call stack visualization</p>
+</body>
+</html>
