+++ media/visual.css
@@ -1,0 +1,87 @@
+html, body{
+    padding:0;
+    margin: 0;
+    box-sizing: border-box;
+    font-family: monospace;
+    }
+#container{
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    background-color: #333;
+    width: 50%;
+    height: 90vh;
+    margin: 2.5% auto;
+    }
+
+.block-attrib{
+    width: 100%;
+    text-align: center;
+    border-bottom: 1px dashed black;
+    font-size: 1.2em;
+    color: #004AB5;
+}
+
+#stack{
+    border-top: 10px solid #2F4858;
+    border-right: 10px solid #2F4858;
+    border-left: 10px solid #2F4858;
+    transform:scaleY(-1);
+    overflow-x: auto; 
+    overflow-y: none;
+    height: 400px;
+    width: 70%;
+    margin: auto;
+}
+
+.title{
+    color: rgb(231, 100, 52);
+    border-bottom: 1px solid red;
+    font-size: 1.5em;
+    text-align: center;
+    padding: 2%;
+}
+
+.block{
+    background-color: #00C896;
+    width: 90%;
+    min-height: 15%;
+    margin: 1% auto;
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;	
+    transform: scaleY(-1);
+}
+    .center{
+    text-align: center;
+    }
+    ::-webkit-scrollbar{
+    width: 6px;
+    }
+    
+    ::-webkit-scrollbar-thumb{
+    border-radius: 10px;
+    background-color: #eee;
+    height: 5px;
+    }
+    
+    ::-webkit-scrollbar-thumb:window-inactive{
+    background: rgba(100,100,100,0.4);
+    }
+    
+    .block-attrib:hover{
+    background-color: #06ebb2;
+    cursor: pointer;
+    }
+
+    .tree-node {
+    border: 1px solid #fff;
+    height: 200px;
+    width: 200px;
+    background-color: #242f38;
+    border-radius: 50%;
+    margin-bottom: 200px;
+    text-align: center;
+    vertical-align: middle;
+    }
